requirements.txt
Django==5.2.10
djangorestframework
djoser
django-cors-headers
django-extensions
python-dotenv
Pillow
